utils/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "reportlab>=4.4.4",
    "spacy>=3.8.7",
    "werkzeug>=3.1.3",
]